requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
requests
ipykernel
pandas
numpy
openpyxl
pyarrow
